refactor: :recycle: files/folder cleanup
