+++ src/pages/app/action_group/index.tsx
@@ -1,192 +1,161 @@
 import React, { useEffect, useState } from "react";
-import { Matrix } from "components/matrix";
+import { Space, Table, Button, Popconfirm } from "antd";
 import type { ColumnsType } from "antd/es/table";
-import { Select, Input, Button } from "antd";
+import { useNavigate, useParams } from "react-router-dom";
 import { Service } from "service";
 import { Endpoint } from "service/endpoint";
-import { useParams } from "react-router-dom";
-import { ActionKind } from "constant/action_kind";
-import { Targets } from "constant/target";
-import { v4 as uuidv4 } from "uuid";
+import { CreateModal } from "components/create_modal";
+import { DeleteOutlined, EditOutlined, PlusOutlined } from "@ant-design/icons";
 import { PageHeader } from "components/page_header";
-import { PlusOutlined } from "@ant-design/icons";
-import "./style.css";
 
-export const Action: React.FC = () => {
+interface DataType {
+  key: string;
+  name: string;
+  description: string;
+  createdAt: string;
+  createdBy: string;
+}
+
+export const ActionGroupDashboard: React.FC = () => {
+  const navigate = useNavigate();
   const [dataSource, setDataSource] = useState([] as any);
-  const [updateData, setupdateData] = useState({} as any);
-  const [savedData, setSavedData] = useState({ data_kind: "Static" } as object);
-  const { appId = "", actionGroupId = "" } = useParams();
-
-  interface DataType {
-    key: string;
-    id: string;
-    description: string;
-    execution_order: string;
-    kind: string;
-    target_value: string;
-    target_kind: string;
-    data_value: string;
-    data_kind: string;
-  }
-
-  /**
-   * fetchActions - will get all the Action for specific action group
-   */
-  const fetchActions = async () => {
-    console.log(appId, "    - ", actionGroupId);
-    await Service.get(`${Endpoint.v1.action.list(appId, actionGroupId)}`)
-      .then((actions) => {
-        setDataSource(actions);
-      })
-      .finally(() => {
-        // all the fallback code will come here
-      });
-  };
-
-  const addNewRow = async () => {
-    let _uuid = uuidv4();
-    let dataSourceLength = dataSource.length + 1;
-    let action = {
-      id: _uuid,
-      execution_order: dataSourceLength,
-      description: "",
-      data_kind: null,
-      data_value: null,
-      target_kind: null,
-      target_value: null,
-      action_group_id: actionGroupId
-    };
-    setDataSource([...dataSource, action]);
-    console.log(dataSource);
-  };
-
-  useEffect(() => {
-    fetchActions();
-  }, []);
+  const [isCreateModalOpen, setIsCreateModalOpen] = useState(false);
 
   const columns: ColumnsType<DataType> = [
     {
-      title: "Command",
-      dataIndex: "command",
-      key: "kind",
+      title: "Name",
+      dataIndex: "name",
+      key: "name",
       render: (text, record) => (
-        <Select
-          defaultValue={record.kind}
-          onChange={(value: any, o: any) =>
-            handleRowUpdate(record, "kind", value)
-          }
-          id="command_select"
-          key="command_select"
-          options={ActionKind.map((x) => {
-            return { key: x.value, value: x.value, label: x.value };
-          })}
-        />
+        <Button type="link" onClick={() => onHandleClick(record)}>
+          {text}
+        </Button>
       )
     },
     {
-      title: "Target",
-      dataIndex: "target",
-      key: "target",
-      render: (text, record) => (
-        <div className="targetContainer">
-          <Select
-            onChange={(value: any, o: any) =>
-              handleRowUpdate(record, "target_kind", value)
-            }
-            options={Targets}
-            defaultValue={record.target_kind}
-          />
-          <Input
-            placeholder="Please enter target"
-            onChange={(e: any) =>
-              handleRowUpdate(record, "target_value", e.target.value)
-            }
-            id="target_value"
-            defaultValue={record.target_value}
-          />
-        </div>
-      )
+      title: "Description",
+      dataIndex: "description",
+      key: "description"
     },
+    // {
+    //   title: "CreatedAt",
+    //   dataIndex: "createdAt",
+    //   key: "createdAt"
+    // },
+    // {
+    //   title: "CreatedBy",
+    //   key: "createdBy",
+    //   dataIndex: "createdBy"
+    // },
     {
-      title: "Value",
-      dataIndex: "value",
-      key: "data_value",
-      render: (text, record) => (
-        <Input
-          placeholder="Please enter value"
-          onChange={(e: any) =>
-            handleRowUpdate(record, "data_value", e.target.value)
-          }
-          id="data_value"
-          defaultValue={record.data_value}
-        />
-      )
+      title: "Action",
+      key: "action",
+      render: (record) => {
+        return (
+          <Space size="middle">
+            <Button
+              type="primary"
+              onClick={() => onHandleClick(record)}
+              shape="round"
+              icon={<EditOutlined />}
+              size="small"
+            />
+            <Popconfirm
+              title="Delete the Action Group"
+              description="Are you sure to delete this Action Group?"
+              onConfirm={() => onDeleteActionGroup(record.id)}
+              okText="Yes"
+              cancelText="No"
+            >
+              <Button
+                danger
+                type="primary"
+                shape="round"
+                icon={<DeleteOutlined />}
+                size="small"
+              />
+            </Popconfirm>
+          </Space>
+        );
+      }
     }
   ];
 
-  const handleRowUpdate = (row: any, field: string, value: string) => {
-    console.log(row, value, field);
-    if (updateData[row.id] !== undefined) {
-      row = updateData[row.id];
-    }
-    row[field] = value;
-    let up: any = {};
-    up[row.id] = row;
+  const { appId = "" } = useParams();
 
-    setupdateData(up);
+  useEffect(() => {
+    fetchActionGroups();
+  }, []);
+
+  /**
+   * onHandleClick - Handle the Action redirect
+   * @param record
+   */
+  const onHandleClick = (record: any) => {
+    navigate(record.id);
   };
 
-  const handleAddRows = () => {
-    setDataSource([
-      ...dataSource,
-      {
-        key: "1",
-        command: "click",
-        target: "Kissflow First Action Group",
-        value: "on click data need to load"
-      }
-    ]);
+  /**
+   * fetchActionGroups - fetch all Action group from the specify Application
+   */
+  const fetchActionGroups = async () => {
+    await Service.get(`${Endpoint.v1.group.getList(appId)}`)
+      .then((actionGroups) => {
+        setDataSource(actionGroups);
+      })
+      .finally(() => {});
   };
 
-  const onHandleInputChange = (event: any) => {
-    setSavedData({ ...savedData, [event.target.id]: event.target.value });
-  };
-
-  const onAddAction = async () => {
-    console.log(dataSource);
-    console.log(updateData);
-
-    await Service.post(`${Endpoint.v1.action.create(appId, actionGroupId)}`, {
-      body: dataSource
+  /**
+   * onAddNewActionGroup - will create new Action Group and
+   * Update the existing grid of all the action group
+   * @param data
+   */
+  const onAddNewActionGroup = async (data: any) => {
+    let payload = { ...data, type_field: "ActionGroup", app_id: appId };
+    await Service.post(`${Endpoint.v1.group.create(appId)}`, {
+      body: payload
     })
-      .then(() => {})
+      .then(() => {
+        fetchActionGroups();
+      })
+      .finally(() => {});
+  };
+
+  /**
+   * onDeleteActionGroup - Delete the Action Group with a confirmation
+   * @param groupId
+   */
+  const onDeleteActionGroup = async (groupId: any) => {
+    await Service.delete(`${Endpoint.v1.group.delete(appId, groupId)}`)
+      .then(() => {
+        fetchActionGroups();
+      })
       .finally(() => {});
   };
 
   return (
-    <div>
+    <>
       <PageHeader
-        backIcon
-        title={""}
+        title="Action Group"
         extra={
-          <div className="btn-footer">
-            <Button type="primary" onClick={addNewRow}>
-              <PlusOutlined /> Row
-            </Button>
-            <Button type="primary" onClick={onAddAction}>
-              Save
-            </Button>
-          </div>
+          <Button type="primary" onClick={() => setIsCreateModalOpen(true)}>
+            <PlusOutlined />
+          </Button>
         }
       />
-      <Matrix
-        dataSource={dataSource}
-        defaultColumns={columns}
-        onAddColumn={null}
-        onAddRow={null}
-        rowKey="id"
-      />
-    </div>
+      <Table columns={columns} dataSource={dataSource} rowKey="name" />
+      <div>
+        {isCreateModalOpen && (
+          <CreateModal
+            isModalOpen={isCreateModalOpen}
+            onClose={() => setIsCreateModalOpen(false)}
+            onOk={onAddNewActionGroup}
+            isLoading={false}
+            modelFor={"Action Group"}
+          />
+        )}
+      </div>
+    </>
   );
 };
