+++ src/pages/app/dashboard/index.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 export function AppDashboard() {
   return <div>This is app Dashboard</div>;
 }
